// src/schemas/transfer.ts
import { z } from "zod";

export const metrcTransferSchema = z.object({
  Id: z.number(),
  ManifestNumber: z.string(),
  ShipmentLicenseType: z.string(),
  ShipperFacilityLicenseNumber: z.string(),
  ShipperFacilityName: z.string(),
  TransporterFacilityLicenseNumber: z.string().nullable(),
  TransporterFacilityName: z.string().nullable(),
  DriverName: z.string().nullable(),
  DriverOccupationalLicenseNumber: z.string().nullable(),
  DriverVehicleLicenseNumber: z.string().nullable(),
  VehicleMake: z.string().nullable(),
  VehicleModel: z.string().nullable(),
  VehicleLicensePlateNumber: z.string().nullable(),
  DeliveryCount: z.number(),
  ReceivedDeliveryCount: z.number(),
  PackageCount: z.number(),
  ReceivedPackageCount: z.number(),
  ContainsPlantPackage: z.boolean(),
  ContainsProductPackage: z.boolean(),
  ContainsTradeSample: z.boolean(),
  ContainsDonation: z.boolean(),
  ContainsTestingSample: z.boolean(),
  ContainsProductRequiresRemediation: z.boolean(),
  ContainsRemediatedProductPackage: z.boolean(),
  CreatedDateTime: z.string(),
  CreatedByUserName: z.string().nullable(),
  LastModified: z.string(),
  DeliveryId: z.number(),
  RecipientFacilityLicenseNumber: z.string(),
  RecipientFacilityName: z.string(),
  ShipmentTypeName: z.string(),
  ShipmentTransactionType: z.string(),
  EstimatedDepartureDateTime: z.string(),
  ActualDepartureDateTime: z.string().nullable(),
  EstimatedArrivalDateTime: z.string(),
  ActualArrivalDateTime: z.string().nullable(),
  DeliveryPackageCount: z.number(),
  DeliveryReceivedPackageCount: z.number(),
  ReceivedDateTime: z.string().nullable(),
  EstimatedReturnDepartureDateTime: z.string().nullable(),
  ActualReturnDepartureDateTime: z.string().nullable(),
  EstimatedReturnArrivalDateTime: z.string().nullable(),
  ActualReturnArrivalDateTime: z.string().nullable(),
});

export type MetrcTransfer = z.infer<typeof metrcTransferSchema>;